requirements.txt
flask
psycopg2-binary
python-dotenv
pycryptodome
requests
Flask-SocketIO